add poses from earlier prototype
These are all the poses used in the earlier prototype.
They are somewhat tuned and proven to work for running man detection.
However, the shin angle used to be relative to the thigh,
no sure if simply translating to absolute angles works, too.
Also, many parameters are not tuned at all, they just "worked".