direction normalized skeletons
It should not matter whether a dancer faces left or right.
But the x sign of angle is flipped for the two cases.

To make comparisons to poses meaningful, normalize the limb angles.
This means, if the body appears to face west, flip all x-coordinates.

Preserve the original direction info for direction-true rendering.