new 3D error scoring function
Now I'm using the euclidean distance of the two directions in a sphere.
Not sure if this is the best.
But it seems it works better than what I've had before.

Tests should finally all be green. But only after swapping to keypoint
files. I think I mixed them up. But can't be 100% sure.