use 2D projection for pose comparisons
After lots of testing, it seems that the z-coordinate is not good enough
to use for full body 3D tracking.
I think it's mostly due to perspective. Because it's mostly  the
feet which are completely off.
If x and y are measured as seen on video, rather than adjusted for
distance, then the z measure is naturally distorted at the feet.

Now, I could try to adjust for distance but it gets more complex.
The prototype showed that 2D is good enough, so I guess I should haven't
even tried 3D.

Of course, some 3D is necessary. I think I will add a discrete
three-state definition for the order between 2 body parts.
I hope the z values are good enough for this.