refactor singleton state into DanceCollection
The dance collection combines all data that is loaded it at runtime
in a single place. There is still only a single global instance of it, for now.
But in the next step this will change. I will probably use a read-only
reference in trackers, which can contain only a subset of all available steps.