actually respect z-order in svg drawing
Finally, a leg that is in front of another according to the pose definition, gets
rendered correctly. The diagonal running man looks so much better now.

Why on earth are there so many code changes for this?
Well, in SVG, the z-index does not work as I thought. Only the order of appearance in
the document matters to determine the draw order.

As a result, I had to completely rethink the declarative SVG.
New, the `SvgLine` component is a virtual, not adding any html/svg.
Instead it registers itself in a surrounding Svg context, which stores
all svg elements in a list and draws them all in the correct order as
defined by a z property I use.

This also meant I had to change how SVG styles are applied. A simple <g>
no longer works. Instead, all styles must be added to each element.
A new SvgStyle context makes managing this easier.

Finally, the biggest annoyance was how reordering SVG elements means
the tweened animation applied to the X and Y values are broken completely.
Svelte seems to assume that the {#each} loop can reuse the component instances from the previous render and just update the property values.
This looks hilariously funky when the x,y values of one body part smooth over to
those of another body part, while (sometimes) also flipping the color.
It looks lik some mad juggling of its own body parts.

The solution/workaround is to make an unanimated round of updates after the
reordering. Crucially, the x and y must stay at the old values per body part but
change in terms of which values each DOM node sees. Then, an await tick().
Then, the new x,y values can be set with animation back on.
I'm a bit worried about the performance of this. But primarily, I'm just happy
it finally works.