first basic, broken instructor avatar
I want a live stick figure to show how one has to move to match the
selected step. This isn't trivial. I have to detect pose by pose instead of
step by step and I don't want to rely on beat timing for it as usual.

This is the basic framework of the code, sort of working. While it doesn't
crash and appears to be doing roughly what it should, while using it, it doesn't feel like detection works as expected.

Furthermore, I need to add various features for a better user experience.
Starting with a success sounds and an instructor figure that is shown in the
right position and size to match the user and dance body shift.