first implementation of pose hints
When checking for a specific next step, the tracker will now
also try to find out what the dancer is doing wrong. It detects
when the wrong leg is in front (left/right switched) and when
the z-ordering of two limbs is the wrong way around.

The UI side will use this information and show left-right
coloring when it makes sense. For z-ordering, there is no
visual clue, yet.