update avatar lengths
Mainly make the feet longer but also adjust the hip to be
a little wider while at it. The feet need to be longer for moves
with the toes down and heel up. Otherwise, the leg angles
aren't right or the toes are in the air.

A negative consequence is the feet are usually not z-adjusted,
with means even when facing the camera, it won't make the
projection shorter. So, I might need to revisit some poses that
look bad as is and maybe add z values to them.