Clearer live recording display
For one, I decided to not show the video feed. I might want
to show it with opacity 0.25 or something. But for now, I want
to get a feeling for how it would be without any video at all.

To further clear up the live recording display, the instructor is
now always in left-right coloring and the live tracked skeleton is
usually in black. It will update to left-right coloring if the pose
hint suggests it.
This makes more sense than the reverse, which I had so far. My
assumption is that users will remember that blue is right, then
they can quickly see which leg should be where if the instructor
has constant left-right coloring.