better pose direction detection
There was an issue when facing the camera with a slightly
turned shoulder, it could happen that the skeleton was
recorded as west-facing. Then all angles are turned 180°
and front-facing comparisons no longer make sense.

This fixes the issue for the next_pose detection but I will
need something additional/better for freestyle detection.