introduce step source
The step source is a bit like a category or a tag. But I decided
to name it source, since I anticipate that if I need something
more general later, I wouldn't get it correct now.

For now, the source is simply from which file the step was
loaded. Every step has exactly one source.
Later I might want the ability to load custom steps, which
would come from a different file, so that would fit in nicely.

Also, I started using wasm_bindgen string interning.
It's premature optimization for sure. Mostly just to
try the provided feature, so that if I decide later that I need
interning everywhere, I already know how this approach feels
in term of coding.