fix dance duration display
Duration initially are in ms, but what I store is in s.
The conversion was not done when adding to the total.

Also, I changed the formatting since I saw some errors
with the old way. Sometimes, no time was displayed, not sure why.
With this, it works. Additionally it shows only the most significant unit (e.g. 1min instead of 1min 4s) which makes
less of a mess on the screen.