more step detection regression tests
- Running Man
- Reverse Running Man
- Kicking Reverse Running Man
- Happy Feet

While these tests are meant as regression tests, right now
they also act as feature completion checks. Successful
detection is not yet guaranteed, most steps are only
configured to look right and have not been thoroughly tested
to be detected by the tracker. By adding these test, I can see
which tests are ready for users to try and track themselves.

Adding these test has also shown how hard it is to detect
steps accurately when not preselecting the step. Especially
when the start and end is not super clean, there can always be
an extra step detected somewhere, or maybe it turns a normal
RM into a double RM. I will need a good solution for that when
hopefully enabling freestyle detection in the future.