fix jump rendering on unchanged Y-values
Some moves have no changes in the Y-coordinate of one or
more body parts. It turned out that svelte was too smart and
decided that no re-rendering is necessary in those cases. Which
had weird results, where some body parts jumped and others
didn't.

By providing a unique id for each update, the tweening and derived jump is triggered as needed for all body parts.

A side-effect is that now the jump is also shown for the idle
steps at the start, which is a bit weird. But that will need a
separate fix.