fix head animation
For some steps, like Pendulum, the head would not move up
and down as expected. Reactivity was not propagated because
the y coordinate was the same, aside from the jump.

To make it work, I added a dummy field to circle and make sure
to update it every time a new skelton is rendered. This ensure
something in the circle object changes, thus propagating
reactivity. Then I also had to change how the physical svg
takes its props. It has to take the entire object, not just the
single values. Otherwise the dummy value is lost in the last
propagation step.