fix: don't match on wrong direction + positioning in normal tracking
2 fixes, one for the problem mentioned in the previous commit, which
caused camera facing positions to match to sideway looking positions.
The fix is to add an error score penalty for the wong direction.
`test_standing_4` is now green with this change.

The other fix is for using tracking on a normal step. It already goes
into the positioning phase first, but it wasn't implemented properly, yet. By adding the idle positions to the unique step tracker, at least
it doesn't crash no more.