delay animation to be on beat
When dancing, one moves before the beat and is already in position
on the beat. The animation so far started moving on the beat and
the froze until the next beat. This looks unnatural and feels
impossible to follow while listening to the music.

Now the animation is delayed enough that the desired position
will be reached just in time for the next beat, freezing in position
for a while then. This feels already much better. The main problem
now is that the tracker expectations are off by one due to this delay.
But that should be solvable.