end course recording after 10  beats
Time limits on course exercises are key for a guided user experience.
This is just a rudimentary structure for it, with simple stats to display
afterwards.

I also added svelte-store as a Rust dependency. I wanted UI reactivity
for a field in Rust. I think I will want more of it soon. This approach
nicely aligns with my architectural principle to have only display code
in JS and do as much of the business logic as possible in Rust.

Depending on svelte-stores seems fine, it's a tiny library with not too
many unnecessary dependencies. It's MIT licensed and easy to read.
Writing the feature myself wouldn't be too time consuming but I
rather try this first and see how well it works and how much I like the API.