fix video review seeking
The video seeking on click didn't work as expected. Only the first click
on a pose worked, with following clicks doing, somehow, nothing.

Now I changed the seeking to only happen when the value changed,
instead of on every frame. At the same time, I removed the explicit
onSeek call after an update from a click on a pose, since the change
will be detected on the next frame.

With these changes, it all works smoothly on desktop and simulated
mobile FF.