implicit z order when looking to the side
When the dancer looks their right, the camera should see
the left half in front of the right side. So far, this was only
possible to specify explicitly, body part for body part in the
pose definition and it wouldn't work for rotated views.

With this commit, all poses facing the side implicitly order
all parts on the left according to the faced direction.