central beat store
Perfectly aligning multiple components to a single beat is
tricky but also very crucial for a good user experience. Thus,
I decided to make another big refactor and put all the beat
information in a common store that can be accessed from all
svelte components. And the trackers can register themselves
to always get updates on the store, too.