better on-beat detection
First, also take camera input delay into consideration.
While my audio and rendered output can be synchronized
to a specific timestamp and thus happen at the same time,
the camera latency is still a problem. By the time the frame
is available, the real user is already ahead by at least 100ms,
which often means they are already half-way in the next pose.

Second, since the accumulated delays will never be perfectly
predictable, I need to be more lenient for detecting poses on
the beat. Furthermore, sometimes a frame just has a bad skeleton
from mediapipe, where the video shows the position is perfect
but the skeleton is completely off.
To be more resilient for both problems, the tracker now gathers
all skeletons that are roughly on the beat and selects the best.
For now, this is only done to errors. Successful matches are counted
immediately, without waiting for a potentially even better match.