translate="no" for all material symbols
When using automated translation on the website, all symbols
were broken. Translation tools didn't recognize them as code, so
they replaced it with a translation, leading to translated text being
displayed in places where symbols should be.

Using the translate="no"  hint, translation tools should ignore it,
as desired.

I've also tried to summarize repeated patterns in abstractions, namely
the plain symbol as <Symbol> and a button with a symbol and a text
as <Button>. This might break a couple of stylings in the short term.
But long term it should be worth it to be able to change styles and
such hints as in this PR in a abstract place rather than in 27+ files.