more robust data loading
- less prerendering to avoid issues with language selection
- fetch leaderboard in load function rather than in render
- move RON files fetching to universal load function rather than server only

These changes are not necessarily positive for performance. I might
want to revert some of the changes if I find they hurt performance.
But I was debugging hosting problems, where the version behind
a publix  reverse proxy didn't work the same as a locally hosted dev
build. And this way of loading data made it simpler to root cause the
issue to reverse proxy configuration problems. Hence, unless proven
to be less efficient, this seems like the simpler thus better approach.