use smaller model by default
Measurements show that on phones, it really is quite slow with
more than 100ms per frame to skeleton conversion.

With the smaller model that goes down to around 70ms. That
is an important step, since a pose-to-pose time of 227ms (132bpm)
we can see 3 frames instead of only 2. The frame closest to the beat
also moves closer in time.

The downside is, I can notice a clear difference in detection quality
due to the leg angles being quite off when not clearly visible.