introduce pose names aside from id
As I'm preparing the pose + step editor, I need to display
a translated string to the user, which should be different
from the unique identifier.

Since I previously named the id name, there is a bit of a shuffling
around to fix things now.
- Pose files in version 0 can still be parsed
and take the old "name" field as the id.
- Existing tests no longer use name but instead use id now.

Otherwise, it should be a mostly smooth transition.