basic z editor
It "works", as in, the pose file can be edited from the UI.
However, setting the value is not intuitive. Setting another
angle would be much more intuitive, essentially editing the
underlying 3D skeleton. But that is not how thing are set up
right now, hence it will be quite a bit more work to get that done.

For now, this works for me but not for other people to actually use it.