change pace within a course
Along the way, clan up the term "beat".
A beat is what we count for bpms, a subbeat is twice that.
A dance at full-pace has a pose per subbeat.

I used to call both beats ans subbeats  the same in code.
Now it should be fixed in most places.