Make production build work again
The upgrade to svelte5 had nasty hidden consequences.

`vite dev` kept working with wasm code, no problem
But `vite preview` didn't. I thought it would be because of
the bun adapter. But no, it is a known problem with top level await
and WASM modules that seeming nobody has solved, yet.

Changing the entire setup to use `vite-plugin-wasm-esm` was the only
thing of a dozen or so that I tried.

This means, the WASM is no longer a src copied in, but an actual ESM module added as a dependency. I guess that makes sense to do anyway.

For now, also removed the custom adapter, just happy that things work again several  hours.