Try to reduce animation artifacts
The head is not always in line with the rest of the body.
I think it's because the animation frame don't align with the updates.
It's probably just one frame off.

This does not actually fix it. I would have to somehow make all updates
belonging to the same avatar transactional. Not sure how to do that just yet.