animated transition to coach page
Hacking around with crossfade to make an illusion of a moved text from
one page to another.

Font size cannot be CSS animated, so I'm using Tween for it.

Built-in crossfade of svelte almost works for the rest. Except, I don't like the
fading out part. And I'm not able to get the sizing right to avoid weird scaling
visuals. To get around that, I copy-pasted the code from crossfade and changed
the hard coded CSS part to have no opacity or scale transformation.