add documentation for services and auth
Some of this is still in progress but it's a good time to write down how the high-level architecture is intended to work.

This documentation is intended for developers (me) since it is getting a bit too complex to (re)build the mental model for the architecture from the code alone.