poc: API to add member to clubs
Add the functionality to add members to a club from the
frontend through an API. Only admins can add members.
The user has no say in it, they are added without confirmation.

This includes a crude way to query a list of users.
It doesn't include display names, yet, since the backend
doesn't have it in the db right now.