implement user meta info sync
Following the design described in sync.md,
data modifications are buffered on the client
and synced with the backend.

The intention is to make offline modification a normal
case that "just works" in all contexts of the app.

This is a simplified version, having only SET operations.
Anything requiring read-modify patterns won't work, yet.