backend: multiple playlists per club
Each club has a channel on the System account.
Clubs can create unlimited playlists on that channel.

One public playlist is used for the main public feed.

Other playlists are for private video sharing.

Using more than one playlist will make it
easier to organize. But this management
from the frontend is not implemented, yet.