display club logos in list view
To make show at the first load, I decided to read
the channel avatars from the API backend and
provide it with the club info.

If I add caching to those calls, it should reduce the
load on PeerTube. The API server is also the better
place for caching than the frontend server, which
wouldn't know when to invalidate the cache.